fix: maven compile error (you can't use @Test in regular classes)
because junit is only defined as a test compile dependency and Maven
expects unit test in src/test/java and not in the regular src/main/java
directory. besides that I'm still not convinced that using @Test like
that is a good idea since it's not a unit test anymore :) We can keep
the tests but make them executable regulary.