feat: saveAsPNG method added to CoreTexture2D + glGetTexImage to CoreGL
CoreTexture2D now has an additional method: saveAsPNG() which will
execute the newly added method glGetTexImage() to download the content
of the texture from the GPU and then save it as a png with the given
filename to disk.

This was especially useful when debugging render to texture issues since
now you can simply save the content of the texture to disk to see what
exactly the content of the texture is.