feat: Expanded CoreLogger with additional functions
The new functions basically improve the performance in case only a few
parameters are used. When providing only functions with var-args
parameters it always leads to object arrays being generated before the
call of the function. To avoid that those objects get created, the
additional functions were added.

Also a crashing error when preparing messages with arguments was fixed
that happened in case the message did not contain enough placeholders
to replace. If this case happens, a additional meaningful error message
is shown in the log.