refactor: CoreShader - Improved fetching of Matrix Type
The original implementation used key values that got very large values.
These values where too large for the box object cache in the
java.lang.Integer class and this resulted in new objects being created
every time a Matrix Type was queried. Since are within very slim bounds
it is easy to create a array where the values are tightly packed and a
array index based retrieval can be used. This avoids the usage of
boxing objects completely.

The downside to this approach is that the calculation for the key needs
to be changed in case the matrix type constants are changed.