chore: remove de.lessvoid:simpleimageloader dependency from examples
The dependency is not in the central available. Since we want to get
coregl in the center which includes the coregl-utils-exmples project
we need to have all dependencies in the central available.

Since the imageloader was only used in the examples (the geometry-shader
example and the texture-atlas example) I had to remove these two
examples as well. That's a bit unfortunate but ok, I think.

We should eventually consider to include the imageloader into the CoreGL
project directly because it's a nice thing to have there, I guess :)