chore: all checkGLError() calls should respect errorChecksEnabled flag
Some of the checkGLError() calls did not check the errorChecksEnabled flag making
it possible to disable error checking with a call to setErrorChecksEnabled(false) and
then still have "some" of them to issue glGetError() internally.

I've now changed it so that it doesn't matter which checkGLError() variation you call.
If the errorChecking is disabled none of them is going to call glGetError and check
for errors.

Additionally I've disabled error checking by default. If needed users can
enable it again with a call to gl.setErrorChecksEnabled(true).

PS: After running a typical Nifty 2.0 scene for a couple of seconds glGetError was
called 3 million times and each of them return GL_NO_ERROR! This is a major waste
of GL time.

Applications should call glGetError() only once each frame by themself and if actual
errors occur should re-run with error checks enabled.