feat: added support for Java 9 (work in progress)
This adds module-info.java and compiler support for Java 9 using
toolchains.xml. To compile with Java 9 you'll need to have both, JDK 8
and JDK 9 installed and a toolchains.xml in place (see readme for details).

We can still compile it without support for Java 9. In this case we'd
still need a toolchains.xml without JDK 9 configured.

Additionally it was necessary to update EasyMock to version 3.4 and
maven-surefire-plugin to version 2.19.1.

The implementation modules still need a module-info.jar which we'll add
in the next commit.