chore: use java 9 service provider modules
This commit introduces module based service providers for providing
CoreGL implementation modules. This allows choosing the CoreGL
implementation by simply having one of the following modules on the
module path:

- requires com.lessvoid.coregl.lwjglthree
- requires com.lessvoid.coregl.lwjgl
- requires com.lessvoid.coregl.jogl

It's still necessary to set -Djava.library.path to point the runtime to
the unpacked native components.

To support different CoreGL implementation in the examples project the
following Maven profiles have been added to the pom.xml:

- examples-jogl
- examples-lwjgl2
- examples-lwjgl3

Besides that the CoreGLUtil interface has been removed and was added to
the CoreGL interface directly or has been moved to the new
CoreBufferUtil class.