feat: new module coregl-ext and renamed maven modules
All Maven modules have been renamed and the "-util" has been removed.
This will change the names of all artifacts as well. However, it's more
consistent now with the Java 9 module names and it makes more sense to
think of "coregl" as a module instead of "coregl.utils".

Additionally a new module "coregl-ext" has been added. This contains
mainly two helper classes that can save textures as png and take
screenshots. The main reason to move these classes from "coregl" to
"coregl-ext" was the dependency to "java.desktop" which will pull in
java.awt. For projects that don't require saving or screenshots this
module is not required and for these project the overall jdk size (in
form of standard modules) is reduced.