chore: remove CoreGLProvider and fix examples to use profiles for lwjgl2/3
CoreGLProvider has been removed. The only class that needs to be
instantiated depending on the adapter is the CoreGL class.

This means that CoreGLSetup does still exists as an interface but is
supposed to be created by the new coreGLSetup() method on the main
CoreGL interface.

In general all adapter implementations will not bring in the dependencies
to lwjgl or jogl any longer. This is the task of the user module. We
like to be an optional dependency. This will make it easier to exist
as a Java 9 module later (the adapter will have a dependency to lwjgl
or jogl but this is not a transient one).