feat: separate example project for java 9 module support added
We now have a dedicated example project how to use coregl as a java 9
module using ServiceLoader. Including this into the existing example
project was too cumbersome.

Additionally this commit applies the compile twice (for java 8 and 9)
compiler-plugin configuration only to modules that need it. Otherwise
it was not possible to have modules compile only for 8 or for 9 when
that configuration was in the parent pom.