chore: CoreShader methods removed that loaded resources from the ClassLoader
These methods would not work properly with JDK9+ anymore without
opening the packages in the module-info.java where the resources
are stored.

Even though that is possible the method parameters and therefore
the methods were generally unfortunately named. It was unclear
that the "filename" parameter actually referenced a classpath
resource and not an actual file.

Removing the methods is a breaking API change but is reducing
the ambiguousness of the API. The remaining methods to load
CoreShader resources either require a File instance or an
InputStream and make there usage more obvious.